% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBoxplot.R
\name{getBoxplot}
\alias{getBoxplot}
\title{getBoxplot}
\usage{
getBoxplot(df)
}
\arguments{
\item{df}{the data.frame returned by \code{getExpression('TP53','TCGA-LUAD')}}
}
\value{
Boxplot figure
}
\description{
getBoxplot
}
\examples{
library(TCGAExpression)
df=getExpression('TP53','TCGA-LUAD')
getBoxplot(df)
}
