% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExpression.R
\name{getExpression}
\alias{getExpression}
\title{getExpression}
\usage{
getExpression(gene, Project)
}
\arguments{
\item{gene}{Offical Gene Symbol: \code{TP53}}

\item{Project}{TCGA projects: \code{TCGA-LUAD},\code{TCGA-OV}}
}
\value{
Gene's Expression across TCGA-patients
}
\description{
getExpression
}
\examples{
library(TCGAExpression)#'
inputToken("yJ0eXAiOiJKV1QiLCJhbGciOiJI...xNjUzODMyOTAyLjM4NTU3MzksImlzcyI6Inl4d")#update your token
getExpression('TP53','TCGA-LUAD')
}
