% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProjects.R
\name{getProjects}
\alias{getProjects}
\title{Get TCGA Project List}
\usage{
getProjects()
}
\value{
33 TCGA Projects List
}
\description{
Get TCGA Project List
}
\examples{
library(TCGAExpression)
getProjects()
}
