% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputToken.R
\name{inputToken}
\alias{inputToken}
\title{Token}
\usage{
inputToken(tokenValues)
}
\arguments{
\item{tokenValues}{Get your token:  http://yixuetongji.top/TCGAToken/}
}
\value{

}
\description{
Token
}
\examples{
library(TCGAExpression)
inputToken("yJ0eXAiOiJKV1QiLCJhbGciOiJI...xNjUzODMyOTAyLjM4NTU3MzksImlzcyI6Inl4d")
}
